+++ p2panda-rs/src/lib.rs
@@ -55,12 +55,17 @@
     unused_qualifications
 )]
 
+// This must be imported here at the root of the crate in order for the fixture macros to work as expected.
+#[cfg(test)]
+use rstest_reuse;
+
 pub mod entry;
 pub mod hash;
 pub mod identity;
 pub mod message;
 pub mod schema;
 pub mod secret_group;
+pub mod test_utils;
 #[cfg(target_arch = "wasm32")]
 pub mod wasm;