+++ CHANGELOG.md
@@ -9,7 +9,11 @@ Highlights are marked with a pancake 🥞
 
 ## [Unreleased]
 
+### Added
+
 - `SecureGroup` for symmetric and asymmetric group encryption [#130](https://github.com/p2panda/p2panda/pull/130) `rs` 🥞
+- `SchemaBuilder` and `Schema` structs for working with CDDL definitions [#78](https://github.com/p2panda/p2panda/pull/78) `rs`
+- `test_utils` module containing `rstest` fixtures, mock `Node` and `Client` structs, test data helper for `p2panda-js` [#116](https://github.com/p2panda/p2panda/pull/116) `rs`
 
 ## [0.2.1]